@@ -20,17 +20,16 @@
   "private": true,
   "dependencies": {
     "@angular-redux/router": "^10.0.0",
-    "@angular-redux/store": "9",
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
-    "@angular/fire": "^5.1.3",
-    "@angular/forms": "^7.2.2",
-    "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
+    "@angular-redux/store": "10.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/fire": "^5.2.1",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
     "@angular/pwa": "^0.12.4",
-    "@angular/router": "^7.2.2",
-    "@angular/service-worker": "^7.2.2",
+    "@angular/router": "^8.0.0",
+    "@angular/service-worker": "^8.0.0",
     "@ionic-native/core": "5.6.1",
     "@ionic-native/file": "5.6.1",
     "@ionic-native/file-opener": "5.6.1",
@@ -38,7 +37,7 @@
     "@ionic-native/social-sharing": "5.6.1",
     "@ionic-native/splash-screen": "5.6.1",
     "@ionic-native/status-bar": "5.6.1",
-    "@ionic/angular": "^4.1.0",
+    "@ionic/angular": "4.4.2",
     "@ionic/storage": "^2.2.0",
     "@ngx-translate/core": "11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
@@ -57,25 +56,25 @@
     "papaparse": "^4.5.0",
     "pdfjs-dist": "^2.0.489",
     "redux": "^4.0.0",
+    "redux-devtools-extension": "^2.13.8",
     "redux-logger": "^3.0.6",
     "redux-observable": "^1.1.0",
     "rxjs": "~6.5.1",
     "sentry-cordova": "^0.15.0",
-    "survey-angular": "^1.0.89",
+    "survey-angular": "1.0.81",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/architect": "~0.13.8",
-    "@angular-devkit/build-angular": "~0.13.8",
+    "@angular-devkit/build-angular": "~0.800.1",
     "@angular-devkit/core": "~7.3.8",
     "@angular-devkit/schematics": "~7.3.8",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.2",
-    "@angular/language-service": "~7.2.2",
+    "@angular/cli": "~8.0.1",
+    "@angular/compiler": "~8.0.0",
+    "@angular/compiler-cli": "~8.0.0",
+    "@angular/language-service": "~8.0.0",
     "@ionic/angular-toolkit": "~1.5.1",
-    "@ionic/v4-migration-tslint": "^1.7.1",
     "@types/c3": "^0.6.0",
     "@types/file-saver": "^1.3.1",
     "@types/html2canvas": "^0.0.35",
@@ -85,7 +84,7 @@
     "@types/node": "^12.0.3",
     "@types/papaparse": "4.1.34",
     "@types/request": "2.47.0",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "~5.1.0",
     "flux-standard-action": "^2.0.3",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
@@ -99,7 +98,7 @@
     "ts-node": "~8.1.0",
     "tslint": "~5.16.0",
     "tslint-config-prettier": "^1.13.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.5"
   },
   "description": "picsa-extension-toolkit: For use by extension training with PICSA"
 }
