@@ -2,7 +2,7 @@ import { enableProdMode } from "@angular/core";
 import { platformBrowserDynamic } from "@angular/platform-browser-dynamic";
 
 import { AppModule } from "./app/app.module";
-import { ENVIRONMENT } from "./environments/environment";
+import ENVIRONMENT from "./environments/environment";
 
 if (ENVIRONMENT.production) {
   enableProdMode();
