@@ -2,7 +2,7 @@ import { NgRedux } from "@angular-redux/store";
 import { Component, Input } from "@angular/core";
 import { AppState } from "src/app/store/store.model";
 import { BudgetToolActions } from "../../budget-tool.actions";
-import { IBudgetCard } from "../../budget-tool.models";
+import { IBudgetCard, ICustomBudgetCard } from "../../budget-tool.models";
 import REGIONAL_SETTINGS from "src/environments/region";
 
 @Component({
@@ -10,7 +10,8 @@ import REGIONAL_SETTINGS from "src/environments/region";
   templateUrl: "budget-card.html"
 })
 export class BudgetCardComponent {
-  @Input("card") card: IBudgetCard;
+  // use partial as not sure whether will be budget card or custom budget card
+  @Input("card") card: Partial<ICustomBudgetCard>;
   @Input("type") type: string;
   selected: boolean;
   currency = REGIONAL_SETTINGS.currency;
@@ -19,4 +20,9 @@ export class BudgetCardComponent {
     public ngRedux: NgRedux<AppState>,
     public actions: BudgetToolActions
   ) {}
+
+  cardClicked() {
+    // *** TODO - figure out what wanted handler to do
+    console.log("card clicked");
+  }
 }
