@@ -4,6 +4,9 @@ import { TranslateModule } from "@ngx-translate/core";
 import { ClimateChartComponent } from "./climate-chart/climate-chart";
 import { CombinedProbabilityComponent } from "./combined-probability/combined-probability";
 import { SiteSelectComponent } from "./site-select/site-select";
+import { TranslateSharedLazyModuleModule } from "src/app/shared/translate-shared-lazy-module.module";
+import { CommonModule } from "@angular/common";
+import { FormsModule } from "@angular/forms";
 
 @NgModule({
   declarations: [
@@ -11,7 +14,13 @@ import { SiteSelectComponent } from "./site-select/site-select";
     CombinedProbabilityComponent,
     SiteSelectComponent
   ],
-  imports: [IonicModule, TranslateModule.forChild()],
+  imports: [
+    IonicModule,
+    TranslateModule.forChild(),
+    TranslateSharedLazyModuleModule,
+    CommonModule,
+    FormsModule
+  ],
   exports: [
     ClimateChartComponent,
     CombinedProbabilityComponent,
